pyproject.toml
[project]
name = "hda"
version = "0.1.0"
description = "Holistic DNA Analyzer — a Python/agent framework that reads your DNA and answers your questions"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "pandas>=2.2",
    "httpx>=0.27",
    "rich>=13.0",
    "click>=8.1",
    "pypdf>=5.0",
]

[project.optional-dependencies]
export = [
    "reportlab>=4.0",
]
dashboard = [
    "streamlit>=1.40",
    "plotly>=5.24",
]

[project.scripts]
hda = "hda.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hda"]
